+++ src/CraftBulkEdit.php
@@ -10,29 +10,34 @@
 
 namespace kffein\craftbulkedit;
 
-use kffein\craftbulkedit\fields\craftbulkeditField as craftbulkeditFieldField;
+use kffein\craftbulkedit\models\SettingsModel;
+use kffein\craftbulkedit\actions\EditDropdown;
+use kffein\craftbulkedit\assetbundles\craftbulkedit\CraftBulkEditAsset;
 
 use Craft;
 use craft\base\Element;
 use craft\base\Plugin;
 use craft\elements\Entry;
-use craft\services\Plugins;
-use craft\events\PluginEvent;
 use craft\services\Fields;
+use craft\services\Plugins;
+use craft\services\Elements as SeviceElements;
+use craft\events\PluginEvent;
+use craft\events\ElementActionEvent;
 use craft\events\RegisterComponentTypesEvent;
 use craft\events\RegisterElementActionsEvent;
+use craft\web\View;
 
 use yii\base\Event;
 
 /**
- * Class craftbulkedit
+ * Class CraftBulkEdit
  *
  * @author    KFFEIN
  * @package   craftbulkedit
  * @since     1.0.1
  *
  */
-class craftbulkedit extends Plugin
+class CraftBulkEdit extends Plugin
 {
     // Static Properties
     // =========================================================================
@@ -48,7 +53,7 @@ class craftbulkedit extends Plugin
     /**
      * @var string
      */
-    public $schemaVersion = '1.0.1';
+    public $schemaVersion = '1.0.2';
 
     // Public Methods
     // =========================================================================
@@ -60,10 +65,41 @@ class craftbulkedit extends Plugin
     {
         parent::init();
         self::$plugin = $this;
+        
+        Event::on(Entry::class, Element::EVENT_REGISTER_ACTIONS, function(RegisterElementActionsEvent $event){
+            Craft::$app->getView()->registerAssetBundle(CraftBulkEditAsset::class);
+            Craft::$app->getView()->registerJs("new Craft.BulkEditPlugin()");
 
-        Event::on(Entry::class, Element::EVENT_REGISTER_ACTIONS, function(RegisterElementActionsEvent $event) {
+            // Get settings from config/
             $settings = $this->getSettings();
-            var_dump($settings);die;
+
+            // Loop through all first level of the settings which should be section's handle.
+            foreach ($settings->addEditFieldAction as $sectionHandle => $fields) {
+                // Get section by the handle from craft section's service
+                $section = Craft::$app->sections->getSectionByHandle($sectionHandle);
+
+                // Add the element list action only if it's a section and it's a valid existant section
+                if(!empty($section) && strpos($event->source,'section')!==false && strpos($event->source, $section->id) !==false ){
+
+                    // Loop through the settings fields list associated to the section
+                    foreach ($fields as $fieldHandle) {
+
+                        // Get the field object
+                        $field = Craft::$app->fields->getFieldByHandle($fieldHandle);
+                        // Get the field class
+                        $type = get_class($field);
+
+                        // Add an action  by passing argument to his constructor based on his type
+                        switch ($type) {
+                            case 'craft\fields\Dropdown':
+                                $event->actions[] = new EditDropdown($field->name,$field->handle,$field->options);
+                                break;
+                            default:
+                                break;
+                        }
+                    }
+                }
+            }
         });
 
         Craft::info(
@@ -78,5 +114,8 @@ class craftbulkedit extends Plugin
 
     // Protected Methods
     // =========================================================================
-
+    protected function createSettingsModel()
+    {
+        return new SettingsModel();
+    }
 }