+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.0.2 - 2018-08-24
+### Added
+- Dynamic selection of sections and fields to bulk edit in the elements listing
+- The only field valid for now is a dropdown.
+
+### Goal
+- plan to make this plugin enabled for other field like radio-button and category for further version.
+
 ## 1.0.1 - 2018-07-26
 ### Added
 - Initial release